requirements.txt
b24pysdk==0.1.0a1
pydantic>=2.0.0
python-dotenv>=1.0.0
flask>=3.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
requests>=2.31.0

